backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Embeddings
openai==1.10.0
anthropic==0.18.0
cohere==4.47
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-cohere>=0.0.1

# Vector DB
pinecone-client==3.0.3

# Monitoring (Optional)
langsmith>=0.1.0

# Utilities
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
